Refactor: Refactor the codebase into logical partitions, add makefile
