+++ internal/kv/encoding.go
